Moved symmetric? guts to symmetric-matrix-entries?
I'm using the same code in the vectorz-clj versions of symmetric?
Putting this code in a separate function in both core.matrix and
vectorz-clj helps highlight that there is identical code,
and maybe there's a good place to put symmetric-matrix-entries?
in core.matrix so that it can be called from vectorz-clj.  I'm not
sure where that is yet.