Fixes default clamp operation
Refactor check for ordering of min and max arguments to be outside element-map function in element-clamp. Also, added check for ordering of min and max arguments for scalars.

Remove test for error on unordered min & max arguments passed to clamp from API tests - some implementations may not want to implement this. Instead added tests for error on unordered min & max arguments passed to clamp to numbers and persistent-vector implementation tests.